fixes to the microscopy extension and mode
